bucket 4j gateway add socket

+++ member/src/main/java/my/application/member/controller/MemberRestController.java
@@ -2,7 +2,7 @@ package my.application.member.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import my.application.member.entities.mysql.member.MemberEntity;
+import my.application.member.entities.mysql.MemberEntity;
 import my.application.member.services.member.MemberService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,7 +20,7 @@ public class MemberRestController {
     private final MemberService memberService;
 
     @GetMapping("")
-    public MemberEntity getMember(@RequestParam(name = "mem_no") Integer memNo) {
+    public MemberEntity getMember(@RequestParam(name = "mem_no") Long memNo) {
         return memberService.getMember(memNo);
     }
     @GetMapping("/members/permit-all")
