+++ member/src/main/java/my/application/member/repositories/mysql/member/MemberRepository.java
@@ -1,13 +1,13 @@
 package my.application.member.repositories.mysql.member;
 
 
-import my.application.member.entities.mysql.member.MemberEntity;
+import my.application.member.entities.mysql.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberRepository extends JpaRepository<MemberEntity,Integer> {
+public interface MemberRepository extends JpaRepository<MemberEntity,Long> {
 
-    MemberEntity findByMemNo(Integer memNo);
+    MemberEntity findByMemNo(Long memNo);
     MemberEntity findById(String memId);
 }
