+++ member/src/main/java/my/application/member/services/member/MemberService.java
@@ -1,7 +1,7 @@
 package my.application.member.services.member;
 
 import lombok.RequiredArgsConstructor;
-import my.application.member.entities.mysql.member.MemberEntity;
+import my.application.member.entities.mysql.MemberEntity;
 import my.application.member.repositories.mysql.member.MemberRepository;
 import org.springframework.stereotype.Service;
 
@@ -16,7 +16,7 @@ public class MemberService {
         return memberRepository.findAll();
     }
 
-    public MemberEntity getMember(Integer memNo) {
+    public MemberEntity getMember(Long memNo) {
         return memberRepository.findByMemNo(memNo);
     }
     public MemberEntity getMember(String id) {
